requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.10.1
PyMySQL==1.1.1
Werkzeug==3.1.3
SQLAlchemy==2.0.40
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
click==8.2.1
blinker==1.9.0
greenlet==3.2.2
typing-extensions==4.13.2
